% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DacianGheorghe.R
\name{DacianGheorghe}
\alias{DacianGheorghe}
\title{DacianGheorghe_tmp}
\usage{
DacianGheorghe()
}
\description{
\code{?DacianGheorghe} contains notes on various functions.\cr
Last update: 05/09/2023 16:34 \cr
Functions in DacianGheorghe Package
\strong{DacianGheorghe} \cr
\strong{go} \cr
\strong{table_} \cr
\strong{rquotes} \cr
\strong{rmvn} \cr
\strong{hello} \cr
}
