% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquotes.R
\name{rquotes}
\alias{rquotes}
\title{rquotes}
\usage{
rquotes(a)
}
\description{
Converts backslashes in path to foward slashes. You have to add r"()" inside the function. \cr
For Example: \code{"C:\Users\DacianGheorghe\Documents\"} then becomes \code{rquotes(r"(C:\Users\DacianGheorghe\Documents\)")} which in turn returns: \code{"C:/Users/DacianGheorghe/Documents/"}.
}
